delivery_service/requirements.txt
fastapi
uvicorn
httpx
pika
aio_pika
python-dotenv
requests
python-consul